+++ changelog.md
@@ -21,6 +21,7 @@
 * As this was a lot harder to do than anticipated, I still need a bit more time to ensure it's working.
 
 * Added Battle for Azeroth trinkets to the autopass table.
+* Non-tradeable and rejected trades from PL is now registered in the loot history.
 
 
 ###### Bugfixes